+++ src/tests/action.rs
@@ -1,8 +1,10 @@
 // SPDX-License-Identifier: MIT
 
 use crate::{
-    Ieee80211ActionFrame, Ieee80211AuthFrameSae, Ieee80211BtmRequest,
-    Ieee80211BtmResponse, Ieee80211Frame, Ieee80211NeighborReportRequest,
+    Ieee80211ActionFrame, Ieee80211ActionFrameBtmResponse,
+    Ieee80211ActionFrameNeighborReportRequest, Ieee80211ActionFrameOther,
+    Ieee80211AuthFrameSae, Ieee80211BtmRequest, Ieee80211BtmResponse,
+    Ieee80211Frame, Ieee80211NeighborReportRequest,
     Ieee80211NeighborReportResponse, Ieee80211StatusCode,
 };
 
@@ -26,8 +28,10 @@ fn action_frame_other_roundtrip() {
     let sta = [0x02u8; 6];
     let ap = [0x01u8; 6];
     let body = b"action body".to_vec();
-    let raw =
-        Ieee80211ActionFrame::new(sta, ap, 0x7e, 0x01, body.clone()).to_bytes();
+    let frame: Ieee80211ActionFrame =
+        Ieee80211ActionFrameOther::new(sta, ap, 0x7e, 0x01, body.clone())
+            .into();
+    let raw = frame.to_bytes();
 
     let parsed = Ieee80211ActionFrame::parse(&raw).expect("parse action frame");
     let Ieee80211ActionFrame::Other(frame) = &parsed else {
@@ -48,14 +52,15 @@ fn action_frame_other_roundtrip() {
 
 #[test]
 fn action_frame_accepts_empty_body() {
-    let raw = Ieee80211ActionFrame::new(
+    let frame: Ieee80211ActionFrame = Ieee80211ActionFrameOther::new(
         [0x02u8; 6],
         [0x01u8; 6],
         0x7e,
         0x01,
         Vec::new(),
     )
-    .to_bytes();
+    .into();
+    let raw = frame.to_bytes();
     assert_eq!(raw.len(), 26);
 
     let parsed = Ieee80211ActionFrame::parse(&raw).expect("parse action frame");
@@ -64,14 +69,15 @@ fn action_frame_accepts_empty_body() {
 
 #[test]
 fn action_frame_parse_rejects_short_frame() {
-    let mut raw = Ieee80211ActionFrame::new(
+    let frame: Ieee80211ActionFrame = Ieee80211ActionFrameOther::new(
         [0x02u8; 6],
         [0x01u8; 6],
         0x7e,
         0x01,
         Vec::new(),
     )
-    .to_bytes();
+    .into();
+    let mut raw = frame.to_bytes();
     raw.pop();
     assert!(Ieee80211ActionFrame::parse(&raw).is_err());
 }
@@ -93,13 +99,14 @@ fn action_frame_parse_rejects_other_subtypes() {
 fn parse_wnm_btm_request() {
     let sta = [0x02u8; 6];
     let ap = [0x01u8; 6];
-    let mut frame = Ieee80211ActionFrame::new(
+    let mut frame: Ieee80211ActionFrame = Ieee80211ActionFrameOther::new(
         ap,
         ap,
         Ieee80211BtmRequest::CATEGORY,
         Ieee80211BtmRequest::ACTION,
         btm_request_body(7),
-    );
+    )
+    .into();
     // A BTM Request travels AP -> STA.
     if let Ieee80211ActionFrame::Other(other) = &mut frame {
         other.da = sta;
@@ -108,11 +115,11 @@ fn parse_wnm_btm_request() {
     let raw = frame.to_bytes();
 
     let parsed = Ieee80211ActionFrame::parse(&raw).expect("parse BTM request");
-    assert!(parsed.is_btm_request());
-    assert!(!parsed.is_btm_response());
     assert_eq!(parsed.category(), Ieee80211BtmRequest::CATEGORY);
     assert_eq!(parsed.action(), Ieee80211BtmRequest::ACTION);
-    let btm = parsed.btm_request().expect("decode BTM request");
+    let Ieee80211ActionFrame::BtmRequest(btm) = &parsed else {
+        panic!("unexpected variant: {parsed:?}");
+    };
     assert_eq!(btm.sta_mac(), sta);
     assert_eq!(btm.request.dialog_token, 7);
     assert!(btm.request.preferred_candidates);
@@ -134,13 +141,14 @@ fn build_wnm_btm_response() {
         target,
     );
     let expected_body = response.build();
-    let raw = Ieee80211ActionFrame::new_btm_response(sta, ap, response.clone())
-        .to_bytes();
+    let frame: Ieee80211ActionFrame =
+        Ieee80211ActionFrameBtmResponse::new(sta, ap, response.clone()).into();
+    let raw = frame.to_bytes();
 
     let parsed = Ieee80211ActionFrame::parse(&raw).expect("parse BTM response");
-    assert!(parsed.is_btm_response());
-    assert!(!parsed.is_btm_request());
-    let btm = parsed.btm_response().expect("decode BTM response");
+    let Ieee80211ActionFrame::BtmResponse(btm) = &parsed else {
+        panic!("unexpected variant: {parsed:?}");
+    };
     assert_eq!(btm.sta_mac(), sta);
     assert_eq!(btm.da, ap);
     assert_eq!(btm.sa, sta);
@@ -159,13 +167,14 @@ fn parse_rrm_neighbor_report_response() {
     body.extend_from_slice(&[81, 6, 7]); // op class, channel, PHY
     body.extend_from_slice(&[6, 3, 1, 2, 3]); // optional subelement
 
-    let mut frame = Ieee80211ActionFrame::new(
+    let mut frame: Ieee80211ActionFrame = Ieee80211ActionFrameOther::new(
         ap,
         ap,
         Ieee80211NeighborReportResponse::CATEGORY,
         Ieee80211NeighborReportResponse::ACTION,
         body,
-    );
+    )
+    .into();
     // A Neighbor Report Response travels AP -> STA.
     if let Ieee80211ActionFrame::Other(other) = &mut frame {
         other.da = sta;
@@ -175,12 +184,10 @@ fn parse_rrm_neighbor_report_response() {
 
     let parsed = Ieee80211ActionFrame::parse(&raw)
         .expect("parse neighbor report response");
-    assert!(parsed.is_neighbor_report_response());
-    assert!(!parsed.is_neighbor_report_request());
     assert_eq!(parsed.sta_mac(), sta);
-    let response = parsed
-        .neighbor_report_response()
-        .expect("decode neighbor report response");
+    let Ieee80211ActionFrame::NeighborReportResponse(response) = &parsed else {
+        panic!("unexpected variant: {parsed:?}");
+    };
     assert_eq!(response.response.dialog_token, 0x2a);
     assert_eq!(response.response.entries.len(), 1);
     let entry = &response.response.entries[0];
@@ -195,20 +202,20 @@ fn build_rrm_neighbor_report_request() {
     let sta = [0x02u8; 6];
     let ap = [0x01u8; 6];
     let request = Ieee80211NeighborReportRequest::new(0x2a);
-    let raw = Ieee80211ActionFrame::new_neighbor_report_request(
-        sta,
-        ap,
-        request.clone(),
-    )
-    .to_bytes();
+    let frame: Ieee80211ActionFrame =
+        Ieee80211ActionFrameNeighborReportRequest::new(
+            sta,
+            ap,
+            request.clone(),
+        )
+        .into();
+    let raw = frame.to_bytes();
 
     let parsed = Ieee80211ActionFrame::parse(&raw)
         .expect("parse neighbor report request");
-    assert!(parsed.is_neighbor_report_request());
-    assert!(!parsed.is_neighbor_report_response());
-    let frame = parsed
-        .neighbor_report_request()
-        .expect("decode neighbor report request");
+    let Ieee80211ActionFrame::NeighborReportRequest(frame) = &parsed else {
+        panic!("unexpected variant: {parsed:?}");
+    };
     assert_eq!(frame.sta_mac(), sta);
     assert_eq!(frame.da, ap);
     assert_eq!(frame.body(), &[0x2a]);
@@ -233,12 +240,14 @@ fn ieee80211_frame_dispatch() {
         Ieee80211Frame::Auth(_)
     ));
 
-    let action = Ieee80211ActionFrame::new_neighbor_report_request(
-        sta,
-        ap,
-        Ieee80211NeighborReportRequest::new(1),
-    )
-    .to_bytes();
+    let action_frame: Ieee80211ActionFrame =
+        Ieee80211ActionFrameNeighborReportRequest::new(
+            sta,
+            ap,
+            Ieee80211NeighborReportRequest::new(1),
+        )
+        .into();
+    let action = action_frame.to_bytes();
     assert!(matches!(
         Ieee80211Frame::parse(&action).expect("parse action"),
         Ieee80211Frame::Action(_)
